refactor(core): move full_pipeline into module app
