+++ src/urdf2mjcf/core.py
@@ -137,25 +137,6 @@ def populate_sensors(mjcf: Element, sensor_config: Element) -> None:
     mjcf.extend(sensor_config.findall("./sensor"))
 
 
-def full_pipeline(
-    urdf: Element,
-    rospack: RosPack = None,
-    sensor_config: Element = None,
-    mujoco_node: Element = None,
-) -> Element:
-    """Convert URDF object to MJCF"""
-    resolve_ros_uris(urdf, rospack)
-    add_mujoco_node(urdf, mujoco_node)
-
-    mjcf = pass_through_mujoco(urdf)
-
-    if sensor_config is not None:
-        populate_sensors(mjcf, sensor_config)
-        mjcf = pass_through_mujoco(mjcf)
-
-    return mjcf
-
-
 # Copyright (c) 2022 Fraunhofer IPA
 #
 # Redistribution and use in source and binary forms, with or without modification, are permitted
