+++ test/test_urdf2mjcf.py
@@ -1,7 +1,8 @@
 import sys
 from pathlib import Path
 
-from urdf2mjcf.core import _parse_element, tostring, full_pipeline as urdf_to_mjcf
+from urdf2mjcf.core import _parse_element, tostring
+from urdf2mjcf.app import full_pipeline as urdf_to_mjcf
 
 _test_dir = Path(__file__).resolve().parent
 sys.path.append(str(_test_dir))
